requirements.txt
# Smart Shopping Platform - Secure AWS Edition
# Essential dependencies for production deployment

fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
pandas==2.1.4
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
